Add handle Error test and rename helper file
